ADD: It’s better to die than to live in the knowledge that you’ve done something that needs forgiveness.
Harry Potter and the Deathly Hallows: Part 2. Actually, I don't like Harry Potter very much, so I haven't seen the film mentioned here. If, like me, you don't like Harry Potter, here is a link to a funny video that has nothing to do with it https://www.youtube.com/shorts/V_L9FdtOhm8. And if not, it's a commit to delete the player when he dies